+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Added LV2 compilation for Linux.
+- Added A/B toggle for switching plugin state.
 - Fixed preset menu not returning to currently selected preset.
 
 ## [1.1.0] 2020-11-25